Separate pekko route from http server instrumentation
